refactor to use solkit
