+++ js/lib/src/instructions/index.ts
@@ -1,3 +1,4 @@
 export * as Item from "./item";
 export * as Namespace from "./namespace";
 export * as Staking from "./staking";
+export * as Matches from "./matches";
