+++ js/lib/src/utils/connection.ts
@@ -19,7 +19,7 @@ export const CLUSTERS: Cluster[] = [
   },
   {
     name: "localnet",
-    url: "http://localhost:8899"
+    url: "http://localhost:8899",
   },
 ];
 export const DEFAULT_CLUSTER = CLUSTERS[2];
